refactor(config): split the 1374-line config.go by concern
Pure code motion — no behaviour change (full suite green, lint clean):

  naming.go      derived OS-user / systemd-unit / firewall-table names
  egress.go      egress containment schema, defaults, per-agent resolution
  vaultbroker.go agent-vault backend, injection services, validation
  sidecar.go     MCP sidecar schema, listener/port allocation, validation
  extensions.go  marketplaces/plugins/skills/MCP servers + validation
  config.go      core schema, shared validation regexes, Load()
