+++ internal/config/config.go
@@ -7,7 +7,6 @@ import (
 	"io"
 	"os"
 	"regexp"
-	"sort"
 	"strings"
 	"time"
 
@@ -21,13 +20,6 @@ var validName = regexp.MustCompile(`^[a-z][a-z0-9-]{0,30}$`)
 // snowflakeRe matches a Discord snowflake ID: 17–19 decimal digits.
 var snowflakeRe = regexp.MustCompile(`^[0-9]{17,19}$`)
 
-// serviceNameRe matches an agent-vault service slug: 3–64 lowercase
-// alphanumeric/hyphen characters (agent-vault's own constraint).
-var serviceNameRe = regexp.MustCompile(`^[a-z0-9-]{3,64}$`)
-
-// avNameInvalid matches characters not allowed in an agent-vault vault name.
-var avNameInvalid = regexp.MustCompile(`[^a-z0-9-]+`)
-
 // gitEmailInvalid matches ASCII whitespace and control characters — the
 // characters that corrupt the line-delimited files git_email is written
 // into: ~/.ssh/allowed_signers uses a space-delimited principal field (a
@@ -36,16 +28,6 @@ var avNameInvalid = regexp.MustCompile(`[^a-z0-9-]+`)
 // NEL, NBSP) pass, but neither git config nor OpenSSH's allowed_signers
 // parser treats them as line or field breaks.
 var gitEmailInvalid = regexp.MustCompile(`[\s\x00-\x1f\x7f]`)
-
-// AgentVaultName maps a clem/sops vault name to an agent-vault-compatible vault
-// name: lowercased, with any run of characters outside [a-z0-9-] collapsed to a
-// single hyphen. sops vault names may contain '_' (e.g. dev_to) or uppercase,
-// which agent-vault rejects (names are lowercase alphanumeric/hyphen). Used
-// everywhere clem hands a vault name to agent-vault so e.g. dev_to -> dev-to
-// consistently across seed, token mint, service rules, and the agent proxy env.
-func AgentVaultName(sopsVault string) string {
-	return avNameInvalid.ReplaceAllString(strings.ToLower(sopsVault), "-")
-}
 
 // githubLoginRe matches a valid GitHub username per GitHub's own rules.
 var githubLoginRe = regexp.MustCompile(`^[a-zA-Z0-9-]{1,39}$`)
@@ -70,30 +52,6 @@ var agentNameInvalid = regexp.MustCompile("[\\x00-\\x1f\\x7f'\"\\\\$`]")
 // splits arguments — so only this conservative character set is allowed.
 var validBindRe = regexp.MustCompile(`^[0-9A-Za-z./:_-]+$`)
 
-// vaultRefRe matches ${vault:BUCKET.KEY} in MCP server env values.
-var vaultRefRe = regexp.MustCompile(`\$\{vault:([^.}]+)\.([^}]+)\}`)
-
-// reservedMCPNames are MCP server names clem's runner hardcodes (runner.go); a
-// sidecar's tool name must not collide with them or it would shadow/duplicate a
-// builtin in the agent's .mcp.json.
-var reservedMCPNames = map[string]bool{
-	"discord-bot": true, "slack-mcp": true, "social": true, "browser-render": true,
-}
-
-// secretKeyRe matches an env-var-style credential key (what a sidecar's secrets
-// and the vault keys look like).
-var secretKeyRe = regexp.MustCompile(`^[A-Za-z_][A-Za-z0-9_]*$`)
-
-// extensionNameRe allows alphanumeric names with dots, hyphens, and underscores.
-// Rejects semicolons, spaces, backticks, and other shell-special characters.
-var extensionNameRe = regexp.MustCompile(`^[a-zA-Z0-9][a-zA-Z0-9._-]*$`)
-
-// extensionRepoRe requires owner/repo format with safe characters only.
-var extensionRepoRe = regexp.MustCompile(`^[a-zA-Z0-9._-]+/[a-zA-Z0-9._-]+$`)
-
-// commitHashRe accepts hex SHA strings (full or prefix).
-var commitHashRe = regexp.MustCompile(`^[0-9a-fA-F]+$`)
-
 // OperatorConfig identifies the humans who are trusted to issue instructions
 // to agents via Discord or GitHub. Provisioned agents use these IDs in the
 // generated prompt so no operator ID is hardcoded in clem source.
@@ -106,88 +64,6 @@ var commitHashRe = regexp.MustCompile(`^[0-9a-fA-F]+$`)
 type OperatorConfig struct {
 	DiscordIDs   []string `yaml:"discord_ids"`
 	GitHubLogins []string `yaml:"github_logins"`
-}
-
-// MarketplaceConfig declares a Claude Code plugin marketplace to install.
-// source must be "github". commit is optional; when set, provision verifies
-// the cloned HEAD matches before proceeding (supply-chain pin).
-type MarketplaceConfig struct {
-	Name   string `yaml:"name"`
-	Source string `yaml:"source"`
-	Repo   string `yaml:"repo"`
-	Commit string `yaml:"commit"`
-}
-
-// PluginConfig declares a plugin to install from a named marketplace.
-// Accepts the shorthand string form "pluginName@marketplaceName".
-type PluginConfig struct {
-	Name        string `yaml:"name"`
-	Marketplace string `yaml:"marketplace"`
-}
-
-// UnmarshalYAML accepts "name@marketplace" shorthand and the struct form.
-func (p *PluginConfig) UnmarshalYAML(value *yaml.Node) error {
-	if value.Tag == "!!str" {
-		parts := strings.SplitN(value.Value, "@", 2)
-		if len(parts) != 2 || parts[0] == "" || parts[1] == "" {
-			return fmt.Errorf("plugin shorthand must be name@marketplace, got %q", value.Value)
-		}
-		p.Name, p.Marketplace = parts[0], parts[1]
-		return nil
-	}
-	type plain PluginConfig
-	return value.Decode((*plain)(p))
-}
-
-// SkillConfig declares a skill to clone into ~/.claude/skills/<name>/.
-// When path is set, the skill entrypoint is at ~/.claude/skills/<name>/<path>.
-type SkillConfig struct {
-	Name   string `yaml:"name"`
-	Source string `yaml:"source"`
-	Repo   string `yaml:"repo"`
-	Path   string `yaml:"path"`
-}
-
-// MCPServerConfig declares an MCP server to register in ~/.claude/settings.json.
-// Env values may contain ${vault:BUCKET.KEY} refs resolved at provision time.
-// command/args run as the agent OS user; clem.yaml is operator-controlled.
-type MCPServerConfig struct {
-	Name    string            `yaml:"name"`
-	Command string            `yaml:"command"`
-	Args    []string          `yaml:"args"`
-	URL     string            `yaml:"url"`
-	Env     map[string]string `yaml:"env"`
-}
-
-// ExtensionsConfig declares all extensions to install for an agent at provision
-// time. Provision is idempotent: already-installed extensions are no-ops.
-// Removing an entry does not auto-uninstall; re-provision logs a reminder.
-type ExtensionsConfig struct {
-	Marketplaces []MarketplaceConfig `yaml:"marketplaces"`
-	Plugins      []PluginConfig      `yaml:"plugins"`
-	Skills       []SkillConfig       `yaml:"skills"`
-	MCPServers   []MCPServerConfig   `yaml:"mcp_servers"`
-}
-
-// PermissionsConfig declares opt-in deny rules written to the root-owned
-// /etc/claude-code/managed-settings.json at provision time. Deny lists from
-// all agents on the host are merged into one file (managed-settings is
-// host-level, not per-user). Empty or absent block = no clem-injected denies.
-type PermissionsConfig struct {
-	Deny []string `yaml:"deny"`
-}
-
-// ExpandVaultRefs replaces ${vault:BUCKET.KEY} refs in s using the qualified
-// secrets map from vault.DecryptForAgent (keys are "vaultName.keyName").
-// Unresolvable refs are left as-is.
-func ExpandVaultRefs(s string, secrets map[string]string) string {
-	return vaultRefRe.ReplaceAllStringFunc(s, func(match string) string {
-		m := vaultRefRe.FindStringSubmatch(match)
-		if v, ok := secrets[m[1]+"."+m[2]]; ok {
-			return v
-		}
-		return match
-	})
 }
 
 // CavemanLevel controls whether and at what intensity the caveman plugin is
@@ -247,321 +123,6 @@ type Config struct {
 	Vault            VaultBackend           `yaml:"vault"`
 	MCPSidecars      MCPSidecarsConfig      `yaml:"mcp_sidecars"`
 	Agents           map[string]AgentConfig `yaml:"agents"`
-}
-
-// MCPSidecarsConfig configures privileged MCP "sidecars": MCP servers that hold
-// a secret the agent must USE but not READ, run as a dedicated non-agent system
-// user, and are reached by the agent over a loopback HTTP MCP transport. This is
-// the "sidecar" disposition in clem's secret-protection model — for credentials
-// the broker can't rewrite (gateway WebSocket tokens) or that warrant scoped
-// access (e.g. read-only Elasticsearch). A stdio MCP server cannot provide this:
-// it runs as the agent's own UID, so its secret is in the agent's reach.
-//
-// Provision wiring lives in cmd/provision.go (provisionMCPSidecars) and
-// internal/proxy: the clem-mcp system user, the pinned mcp-proxy stdio→HTTP
-// bridge, one systemd loopback listener per subscribed sidecar (upstream
-// secret supplied root-side via EnvironmentFile), a per-port nftables rule
-// restricting each listener to its subscriber UID(s) — the listener
-// Requires= that firewall unit, fail-closed — and the http entry in the
-// runner-generated .mcp.json. See docs/threat-model.md.
-type MCPSidecarsConfig struct {
-	// SystemUser is the dedicated non-login user that runs all sidecar servers.
-	// Default "clem-mcp".
-	SystemUser string `yaml:"system_user"`
-	// BasePort is the first loopback port allocated to sidecar listeners;
-	// subsequent listeners take successive ports deterministically. Default 14500.
-	BasePort int `yaml:"base_port"`
-	// Servers declares the available sidecar MCP servers; agents subscribe by
-	// name via AgentConfig.Sidecars.
-	Servers []SidecarServer `yaml:"servers"`
-}
-
-// SidecarServer is one privileged MCP server definition. The agent sees only a
-// loopback HTTP MCP endpoint, reachable solely from its own UID (enforced by
-// the sidecar nftables rule); the upstream credential lives only in the
-// sidecar process (sourced from Secrets, written to a root-owned env file,
-// never any agent .env).
-type SidecarServer struct {
-	// Name is the sidecar slug (matches validName) and the MCP server name the
-	// agent sees in .mcp.json unless Tool overrides it.
-	Name string `yaml:"name"`
-	// Identity is "shared" (one upstream credential for all subscribers, e.g.
-	// read-only ES) or "per-agent" (one listener per subscribing agent with that
-	// agent's own credential, e.g. each agent's Discord bot token). Default shared.
-	Identity string `yaml:"identity"`
-	// Command + Args are the underlying stdio MCP server the bridge wraps.
-	Command string   `yaml:"command"`
-	Args    []string `yaml:"args"`
-	// Transport is the agent-facing transport. "http" (streamable-HTTP, default;
-	// recommended) or "sse" (deprecated).
-	Transport string `yaml:"transport"`
-	// Secrets are the credential keys this sidecar holds (written to its
-	// root-owned env file); for identity per-agent they are resolved per agent.
-	Secrets []string `yaml:"secrets"`
-	// SecretsVault is the sops/agent-vault vault to pull Secrets from.
-	SecretsVault string `yaml:"secrets_vault"`
-	// Tool overrides the MCP server name the agent sees (default Name).
-	Tool string `yaml:"tool"`
-}
-
-// SystemUserOrDefault returns the sidecar system user, default clem-mcp.
-func (m MCPSidecarsConfig) SystemUserOrDefault() string {
-	if m.SystemUser == "" {
-		return "clem-mcp"
-	}
-	return m.SystemUser
-}
-
-// BasePortOrDefault returns the first sidecar loopback port, default 14500.
-func (m MCPSidecarsConfig) BasePortOrDefault() int {
-	if m.BasePort == 0 {
-		return 14500
-	}
-	return m.BasePort
-}
-
-// IdentityKind normalizes the identity to "shared" (default) or "per-agent".
-func (s SidecarServer) IdentityKind() string {
-	if s.Identity == "per-agent" {
-		return "per-agent"
-	}
-	return "shared"
-}
-
-// TransportKind normalizes the transport to "http" (default) or "sse".
-func (s SidecarServer) TransportKind() string {
-	if s.Transport == "sse" {
-		return "sse"
-	}
-	return "http"
-}
-
-// ToolName returns the MCP server name the agent sees (Tool or Name).
-func (s SidecarServer) ToolName() string {
-	if s.Tool != "" {
-		return s.Tool
-	}
-	return s.Name
-}
-
-// SidecarPort returns the deterministic loopback port for the i-th sidecar
-// listener (allocated upward from base_port). Stable across provisions so
-// re-provision yields minimal nftables/.mcp.json diffs.
-func (m MCPSidecarsConfig) SidecarPort(i int) int { return m.BasePortOrDefault() + i }
-
-// VaultBackend selects how agent secrets are materialized (Phase 2). Default
-// "env" preserves the legacy flow: secrets decrypted from sops are written
-// verbatim into each agent's .env. "agent-vault" routes HTTP-brokerable
-// credentials through an Infisical agent-vault credential proxy so the real
-// secret never reaches the agent — the agent holds only a placeholder and a
-// scoped, inject-only token. sops remains the git-committable source of truth
-// and seeds agent-vault at provision time.
-type VaultBackend struct {
-	// Backend is "env" (default) or "agent-vault".
-	Backend string `yaml:"backend"`
-	// SystemUser is the dedicated non-login user that owns the vault store and
-	// runs agent-vault. Default "clem-vault".
-	SystemUser string `yaml:"system_user"`
-	// Addr is the agent-vault management API the provisioner seeds/mints against.
-	// Default "http://127.0.0.1:14321".
-	Addr string `yaml:"addr"`
-	// ProxyHost is the host:port of agent-vault's TLS-MITM proxy that agents
-	// point HTTPS_PROXY at. Default "127.0.0.1:14322".
-	ProxyHost string `yaml:"proxy_host"`
-	// CACertPath is where agent-vault's CA cert is written for agents to trust
-	// the intercepted TLS. Default "/etc/clem/agent-vault-ca.pem".
-	CACertPath string `yaml:"ca_cert_path"`
-	// Services are the global injection rules. agent-vault matches outbound
-	// requests by host and swaps in the real credential, so a brokered agent
-	// egresses with the real value while its .env holds only a placeholder. Each
-	// service names credential KEYS (token_key etc.); at provision clem applies
-	// every service whose keys an agent brokers into that agent's consolidated
-	// vault. A brokered secret with no matching service egresses as a placeholder.
-	Services []Service `yaml:"services"`
-}
-
-// Service is one agent-vault injection rule: for requests to Host, attach the
-// credential(s) named below using the AuthType scheme. Services are not bound to
-// a vault — clem applies each one into the consolidated vault of every agent
-// that brokers the referenced credential keys. Mirrors `agent-vault vault
-// service add` flags 1:1.
-type Service struct {
-	// Name is the service slug (3-64 lowercase alphanumeric/hyphen chars).
-	Name string `yaml:"name"`
-	// Host is the target: a bare host (api.x.com), one-level wildcard
-	// (*.x.com), or inline path form (slack.com/api/*).
-	Host string `yaml:"host"`
-	// AuthType is bearer | basic | api-key | custom | passthrough.
-	AuthType string `yaml:"auth_type"`
-	// TokenKey is the credential key for bearer auth.
-	TokenKey string `yaml:"token_key"`
-	// UsernameKey / PasswordKey are the credential keys for basic auth.
-	UsernameKey string `yaml:"username_key"`
-	PasswordKey string `yaml:"password_key"`
-	// APIKeyKey is the credential key for api-key auth; APIKeyHeader (default
-	// Authorization) and APIKeyPrefix (e.g. "Bot ") shape the injected header.
-	APIKeyKey    string `yaml:"api_key_key"`
-	APIKeyHeader string `yaml:"api_key_header"`
-	APIKeyPrefix string `yaml:"api_key_prefix"`
-}
-
-// CredentialKeys returns the credential keys this service injects, per auth type.
-func (s Service) CredentialKeys() []string {
-	switch s.AuthType {
-	case "bearer":
-		return []string{s.TokenKey}
-	case "basic":
-		return []string{s.UsernameKey, s.PasswordKey}
-	case "api-key":
-		return []string{s.APIKeyKey}
-	}
-	return nil
-}
-
-// ValidAuthTypes are the agent-vault service auth schemes clem accepts.
-var ValidAuthTypes = map[string]bool{
-	"bearer": true, "basic": true, "api-key": true, "custom": true, "passthrough": true,
-}
-
-// UnbrokerableSecrets are credential keys that agent-vault cannot broker over
-// HTTP and must therefore stay as real values in .env: the Discord gateway
-// token (sent in a post-upgrade WebSocket frame, not an HTTP header) and the
-// SSH/Elasticsearch creds (not HTTP at all). Listing any of these in
-// brokered_secrets is a hard config error.
-var UnbrokerableSecrets = map[string]bool{
-	"DISCORD_TOKEN": true,
-	"SSH_HOST":      true,
-	"SSH_USER":      true,
-	"SSH_KEY_PATH":  true,
-	"ES_USER":       true,
-	"ES_PASSWORD":   true,
-}
-
-// IsAgentVault reports whether the agent-vault backend is selected.
-func (v VaultBackend) IsAgentVault() bool { return v.Backend == "agent-vault" }
-
-// SystemUserOrDefault returns the vault system user, default clem-vault.
-func (v VaultBackend) SystemUserOrDefault() string {
-	if v.SystemUser == "" {
-		return "clem-vault"
-	}
-	return v.SystemUser
-}
-
-// AddrOrDefault returns the management API address, default localhost:14321.
-func (v VaultBackend) AddrOrDefault() string {
-	if v.Addr == "" {
-		return "http://127.0.0.1:14321"
-	}
-	return v.Addr
-}
-
-// ProxyHostOrDefault returns the MITM proxy host:port, default 127.0.0.1:14322.
-func (v VaultBackend) ProxyHostOrDefault() string {
-	if v.ProxyHost == "" {
-		return "127.0.0.1:14322"
-	}
-	return v.ProxyHost
-}
-
-// CACertPathOrDefault returns the CA cert path, default under /etc/clem.
-func (v VaultBackend) CACertPathOrDefault() string {
-	if v.CACertPath == "" {
-		return "/etc/clem/agent-vault-ca.pem"
-	}
-	return v.CACertPath
-}
-
-// EgressConfig configures hard egress containment via pipelock + a per-agent
-// nftables UID firewall. When enabled, each agent's outbound traffic is forced
-// through a single pipelock forward proxy (run as a dedicated non-login system
-// user) on loopback; everything else is rejected by the kernel firewall, so a
-// compromised agent cannot reach the network except via the proxy.
-//
-// This supersedes the per-agent EgressRestrictionExperimental flag, which used
-// systemd IPAddressAllow with hardcoded CIDRs. The CIDR approach is replaced by
-// domain allowlisting in pipelock + a loopback-only systemd block as a second
-// kernel layer.
-type EgressConfig struct {
-	// Enabled turns on pipelock + the per-agent UID firewall for every agent
-	// that does not individually override via AgentConfig.Egress.
-	Enabled bool `yaml:"enabled"`
-	// Posture maps to pipelock's mode: "strict" (allowlist-only), "balanced"
-	// (block known-bad, default), or "audit" (log-only, block nothing).
-	Posture string `yaml:"posture"`
-	// Domains is the outbound allowlist written into pipelock's api_allowlist.
-	// Hostnames, no scheme. pipelock wildcard "*.example.com" also matches the
-	// apex. Empty falls back to DefaultEgressDomains.
-	Domains []string `yaml:"domains"`
-	// ProxyPort is the loopback port pipelock's forward proxy listens on.
-	// Default 8888.
-	ProxyPort int `yaml:"proxy_port"`
-	// ProxyUser is the dedicated non-login system user that runs pipelock.
-	// Default "clem-proxy".
-	ProxyUser string `yaml:"proxy_user"`
-	// TLSIntercept enables pipelock body/response scanning, which requires
-	// distributing the pipelock CA into each agent's trust stores. Default
-	// false — CONNECT-only (SNI/host allowlist + audit) needs no CA.
-	TLSIntercept bool `yaml:"tls_intercept"`
-	// AllowLocalhostPorts are loopback ports the agent UID may reach besides
-	// the proxy (e.g. 11434 for Ollama). The proxy port is always allowed.
-	//
-	// WARNING: any daemon listening on an allowed loopback port runs as a
-	// non-contained UID and egresses freely, so it is an SSRF pivot — only
-	// list services that cannot be coerced into making outbound requests on
-	// the agent's behalf.
-	AllowLocalhostPorts []int `yaml:"allow_localhost_ports"`
-}
-
-// DefaultEgressDomains is the allowlist applied when egress is enabled but no
-// domains are configured: the minimum an agent needs to reach Anthropic and
-// GitHub. pipelock wildcards match the apex too.
-var DefaultEgressDomains = []string{"*.anthropic.com", "github.com", "*.githubusercontent.com"}
-
-// PostureOrDefault returns the configured pipelock mode, defaulting to balanced.
-func (e EgressConfig) PostureOrDefault() string {
-	if e.Posture == "" {
-		return "balanced"
-	}
-	return e.Posture
-}
-
-// ProxyPortOrDefault returns the configured proxy port, defaulting to 8888.
-func (e EgressConfig) ProxyPortOrDefault() int {
-	if e.ProxyPort == 0 {
-		return 8888
-	}
-	return e.ProxyPort
-}
-
-// ProxyUserOrDefault returns the configured proxy system user, default clem-proxy.
-func (e EgressConfig) ProxyUserOrDefault() string {
-	if e.ProxyUser == "" {
-		return "clem-proxy"
-	}
-	return e.ProxyUser
-}
-
-// DomainsOrDefault returns the configured allowlist or DefaultEgressDomains.
-func (e EgressConfig) DomainsOrDefault() []string {
-	if len(e.Domains) == 0 {
-		return DefaultEgressDomains
-	}
-	return e.Domains
-}
-
-// EgressEnabledFor reports whether egress containment applies to an agent.
-// Resolution order: explicit per-agent override, then the deprecated per-agent
-// egress_restriction_experimental flag, then the top-level egress.enabled.
-func (c *Config) EgressEnabledFor(agentKey string) bool {
-	ac := c.Agents[agentKey]
-	if ac.Egress != nil {
-		return *ac.Egress
-	}
-	if ac.EgressRestrictionExperimental {
-		return true
-	}
-	return c.Egress.Enabled
 }
 
 type Coordination struct {
@@ -812,139 +373,6 @@ func (ac AgentConfig) ProviderEnv() (map[string]string, error) {
 	}
 }
 
-// OSUsername returns the OS username for an agent: <project>-<agentkey>
-func (c *Config) OSUsername(agentKey string) string {
-	return fmt.Sprintf("%s-%s", c.Project, agentKey)
-}
-
-// ServiceName returns the systemd service name for an agent.
-func (c *Config) ServiceName(agentKey string) string {
-	return fmt.Sprintf("clem-%s-%s.service", c.Project, agentKey)
-}
-
-// WatchdogServiceName returns the systemd service name for the watchdog.
-func (c *Config) WatchdogServiceName() string {
-	return fmt.Sprintf("clem-watchdog-%s.service", c.Project)
-}
-
-// WatchdogTimerName returns the systemd timer name for the watchdog.
-func (c *Config) WatchdogTimerName() string {
-	return fmt.Sprintf("clem-watchdog-%s.timer", c.Project)
-}
-
-// TtydServiceName returns the systemd service name for the agent's web terminal.
-func (c *Config) TtydServiceName(agentKey string) string {
-	return fmt.Sprintf("clem-ttyd-%s-%s.service", c.Project, agentKey)
-}
-
-// PipelockServiceName returns the systemd service name for the egress proxy.
-func (c *Config) PipelockServiceName() string {
-	return fmt.Sprintf("clem-pipelock-%s.service", c.Project)
-}
-
-// NftablesServiceName returns the systemd service name for the egress firewall.
-func (c *Config) NftablesServiceName() string {
-	return fmt.Sprintf("clem-nftables-%s.service", c.Project)
-}
-
-// AgentVaultServiceName returns the systemd service name for the agent-vault
-// credential proxy.
-func (c *Config) AgentVaultServiceName() string {
-	return fmt.Sprintf("clem-agent-vault-%s.service", c.Project)
-}
-
-// SidecarServiceName returns the systemd service name for a sidecar listener.
-// For a per-agent sidecar there is one listener per subscriber, so the agent
-// key disambiguates; shared sidecars pass an empty agentKey.
-func (c *Config) SidecarServiceName(name, agentKey string) string {
-	if agentKey != "" {
-		return fmt.Sprintf("clem-mcp-%s-%s-%s.service", c.Project, name, agentKey)
-	}
-	return fmt.Sprintf("clem-mcp-%s-%s.service", c.Project, name)
-}
-
-// SidecarNftablesServiceName returns the systemd service name for the sidecar
-// loopback firewall (distinct from the egress firewall; sidecar isolation must
-// hold even on hosts with no egress containment).
-func (c *Config) SidecarNftablesServiceName() string {
-	return fmt.Sprintf("clem-sidecar-nft-%s.service", c.Project)
-}
-
-// SidecarListener is one concrete loopback listener clem provisions for a
-// privileged MCP sidecar. A shared sidecar yields one listener (Subscribers =
-// every subscribing agent, AgentKey = ""); a per-agent sidecar yields one
-// listener per subscriber (Subscribers = [that agent], AgentKey = that agent).
-type SidecarListener struct {
-	Server      SidecarServer
-	Port        int
-	Subscribers []string // sorted agent keys allowed to reach this listener
-	AgentKey    string   // subscriber key for per-agent identity; "" when shared
-}
-
-// SidecarListeners returns the deterministic listener layout for all subscribed
-// sidecars: the single source of truth for port allocation shared by config
-// validation, the systemd/nftables generators, provisioning, and the runner.
-// Servers are taken in config order; per-agent subscribers in sorted order;
-// ports allocated upward from base_port. Unsubscribed servers yield no listener.
-func (c *Config) SidecarListeners() []SidecarListener {
-	var out []SidecarListener
-	idx := 0
-	for _, s := range c.MCPSidecars.Servers {
-		subs := c.sidecarSubscribers(s.Name)
-		if len(subs) == 0 {
-			continue
-		}
-		if s.IdentityKind() == "per-agent" {
-			for _, ak := range subs {
-				out = append(out, SidecarListener{
-					Server:      s,
-					Port:        c.MCPSidecars.SidecarPort(idx),
-					Subscribers: []string{ak},
-					AgentKey:    ak,
-				})
-				idx++
-			}
-			continue
-		}
-		out = append(out, SidecarListener{
-			Server:      s,
-			Port:        c.MCPSidecars.SidecarPort(idx),
-			Subscribers: subs,
-		})
-		idx++
-	}
-	return out
-}
-
-// sidecarSubscribers returns the sorted agent keys subscribing to a sidecar.
-func (c *Config) sidecarSubscribers(name string) []string {
-	var subs []string
-	for ak, ac := range c.Agents {
-		for _, n := range ac.Sidecars {
-			if n == name {
-				subs = append(subs, ak)
-				break
-			}
-		}
-	}
-	sort.Strings(subs)
-	return subs
-}
-
-// IsBrokered reports whether a secret key is HTTP-brokered for this agent
-// (placeholder in .env, real value only inside agent-vault).
-func (ac AgentConfig) IsBrokered(key string) bool {
-	if !ac.VaultBroker {
-		return false
-	}
-	for _, k := range ac.BrokeredSecrets {
-		if k == key {
-			return true
-		}
-	}
-	return false
-}
-
 // envVarRe matches ${VAR} or ${VAR:-default} for env interpolation in clem.yaml.
 // Names are conservative on purpose: [A-Z_][A-Z0-9_]* — no silent expansion of
 // arbitrary shell constructs.
@@ -1115,201 +543,6 @@ func Load(path string) (*Config, error) {
 	return &cfg, nil
 }
 
-// validateMCPSidecars checks the privileged-sidecar declarations and that every
-// agent subscription references a defined server.
-func (cfg *Config) validateMCPSidecars() error {
-	subscribed := false
-	for _, ac := range cfg.Agents {
-		if len(ac.Sidecars) > 0 {
-			subscribed = true
-		}
-	}
-	if len(cfg.MCPSidecars.Servers) == 0 && !subscribed {
-		return nil
-	}
-	if p := cfg.MCPSidecars.BasePort; p != 0 && (p < 1 || p > 65535) {
-		return fmt.Errorf("mcp_sidecars.base_port %d out of range 1-65535", p)
-	}
-	defined := map[string]bool{}
-	for i, s := range cfg.MCPSidecars.Servers {
-		where := fmt.Sprintf("mcp_sidecars.servers[%d]", i)
-		if s.Name != "" {
-			where = "mcp_sidecars " + s.Name
-		}
-		if !validName.MatchString(s.Name) {
-			return fmt.Errorf("%s: name must match %s", where, validName.String())
-		}
-		if defined[s.Name] {
-			return fmt.Errorf("%s: duplicate sidecar name %q", where, s.Name)
-		}
-		defined[s.Name] = true
-		if s.Command == "" {
-			return fmt.Errorf("%s: command is required", where)
-		}
-		if len(s.Secrets) == 0 {
-			return fmt.Errorf("%s: at least one secret is required (a sidecar with no secret to hide should be a normal MCP server)", where)
-		}
-		if !strings.HasPrefix(s.Command, "/") {
-			return fmt.Errorf("%s: command must be an absolute path, got %q", where, s.Command)
-		}
-		tn := s.ToolName()
-		if !validName.MatchString(tn) {
-			return fmt.Errorf("%s: tool name %q must match %s (set `tool:` to override)", where, tn, validName.String())
-		}
-		if reservedMCPNames[tn] {
-			return fmt.Errorf("%s: tool name %q collides with a builtin MCP server (%v); set a distinct `tool:`", where, tn, []string{"discord-bot", "slack-mcp", "social", "browser-render"})
-		}
-		for _, k := range s.Secrets {
-			if !secretKeyRe.MatchString(k) {
-				return fmt.Errorf("%s: secret key %q must match %s", where, k, secretKeyRe.String())
-			}
-		}
-		switch s.Identity {
-		case "", "shared", "per-agent":
-		default:
-			return fmt.Errorf("%s: identity must be shared or per-agent, got %q", where, s.Identity)
-		}
-		switch s.Transport {
-		case "", "http":
-		case "sse":
-			fmt.Fprintf(os.Stderr, "warning: %s: transport sse is deprecated; prefer http (streamable-HTTP)\n", where)
-		default:
-			return fmt.Errorf("%s: transport must be http or sse, got %q", where, s.Transport)
-		}
-		// A shared sidecar's one credential set comes from a named sops vault;
-		// per-agent reads each subscriber's own vaults at provision, so it needs none.
-		if s.IdentityKind() == "shared" && s.SecretsVault == "" {
-			return fmt.Errorf("%s: a shared sidecar requires secrets_vault (the sops vault to read its secrets from)", where)
-		}
-	}
-	for key, ac := range cfg.Agents {
-		for _, name := range ac.Sidecars {
-			if !defined[name] {
-				return fmt.Errorf("agent %s: sidecar %q is not defined in mcp_sidecars.servers", key, name)
-			}
-		}
-		// The runner only emits the http MCP entry into the claude-code .mcp.json;
-		// an opencode agent would get the listener + firewall but never the config,
-		// a silent no-op. Reject rather than mislead.
-		if len(ac.Sidecars) > 0 && ac.RuntimeKind() == "opencode" {
-			return fmt.Errorf("agent %s: sidecars are not supported with runtime opencode", key)
-		}
-	}
-	// Deterministic loopback-port allocation (SidecarListeners is the single
-	// source of truth). A shared sidecar listens once; a per-agent sidecar once
-	// per subscribing agent. Check the range fits below 65535 and never collides
-	// with another loopback port clem binds on the same host.
-	listeners := cfg.SidecarListeners()
-	base := cfg.MCPSidecars.BasePortOrDefault()
-	if n := len(listeners); n > 0 && base+n-1 > 65535 {
-		return fmt.Errorf("mcp_sidecars: %d listeners from base_port %d overflow past 65535", n, base)
-	}
-	reserved := map[int]string{}
-	for key, ac := range cfg.Agents {
-		if ac.WebTerminalPort != 0 {
-			reserved[ac.WebTerminalPort] = "agent " + key + " web_terminal_port"
-		}
-	}
-	egressInUse := cfg.Egress.Enabled
-	for key := range cfg.Agents {
-		if cfg.EgressEnabledFor(key) {
-			egressInUse = true
-		}
-	}
-	if egressInUse {
-		reserved[cfg.Egress.ProxyPortOrDefault()] = "egress proxy_port"
-		for _, p := range cfg.Egress.AllowLocalhostPorts {
-			reserved[p] = "egress allow_localhost_ports"
-		}
-	}
-	if cfg.Vault.IsAgentVault() {
-		// Default mgmt/MITM ports; a custom vault address near base_port is not parsed here.
-		reserved[14321] = "agent-vault management port"
-		reserved[14322] = "agent-vault MITM port"
-	}
-	for _, l := range listeners {
-		if owner, clash := reserved[l.Port]; clash {
-			return fmt.Errorf("mcp_sidecars: sidecar %q port %d collides with %s", l.Server.Name, l.Port, owner)
-		}
-	}
-	return nil
-}
-
-// validateVaultServices checks the agent-vault injection rules. Services are
-// global (not vault-bound): clem applies each into the consolidated vault of any
-// agent that brokers the referenced credential keys. Warns on services no agent
-// can use and on brokered secrets with no service to inject them.
-func (cfg *Config) validateVaultServices() error {
-	if len(cfg.Vault.Services) == 0 {
-		return nil
-	}
-	if !cfg.Vault.IsAgentVault() {
-		return fmt.Errorf("vault.services requires vault.backend: agent-vault")
-	}
-	// All credential keys brokered by some agent.
-	allBrokered := map[string]bool{}
-	for _, ac := range cfg.Agents {
-		if !ac.VaultBroker {
-			continue
-		}
-		for _, s := range ac.BrokeredSecrets {
-			allBrokered[s] = true
-		}
-	}
-	serviceKeys := map[string]bool{}
-	for i, s := range cfg.Vault.Services {
-		where := fmt.Sprintf("vault.services[%d]", i)
-		if s.Name != "" {
-			where = "vault.services " + s.Name
-		}
-		if !serviceNameRe.MatchString(s.Name) {
-			return fmt.Errorf("%s: name must be 3-64 lowercase alphanumeric/hyphen chars", where)
-		}
-		if s.Host == "" {
-			return fmt.Errorf("%s: host is required", where)
-		}
-		if !ValidAuthTypes[s.AuthType] {
-			return fmt.Errorf("%s: auth_type must be bearer, basic, api-key, custom, or passthrough, got %q", where, s.AuthType)
-		}
-		switch s.AuthType {
-		case "bearer":
-			if s.TokenKey == "" {
-				return fmt.Errorf("%s: auth_type bearer requires token_key", where)
-			}
-		case "basic":
-			if s.UsernameKey == "" || s.PasswordKey == "" {
-				return fmt.Errorf("%s: auth_type basic requires username_key and password_key", where)
-			}
-		case "api-key":
-			if s.APIKeyKey == "" {
-				return fmt.Errorf("%s: auth_type api-key requires api_key_key", where)
-			}
-		}
-		anyBrokered := false
-		for _, k := range s.CredentialKeys() {
-			serviceKeys[k] = true
-			if allBrokered[k] {
-				anyBrokered = true
-			}
-		}
-		if len(allBrokered) > 0 && !anyBrokered {
-			fmt.Fprintf(os.Stderr, "warning: %s references credential keys no agent brokers — it will not be applied\n", where)
-		}
-	}
-	// Warn on any brokered secret with no service to inject it.
-	for key, ac := range cfg.Agents {
-		if !ac.VaultBroker {
-			continue
-		}
-		for _, s := range ac.BrokeredSecrets {
-			if !serviceKeys[s] {
-				fmt.Fprintf(os.Stderr, "warning: agent %s: brokered secret %q has no matching vault.service — it would egress as a placeholder\n", key, s)
-			}
-		}
-	}
-	return nil
-}
-
 // validate checks that all discord_ids and github_logins are well-formed.
 func (op *OperatorConfig) validate() error {
 	for _, id := range op.DiscordIDs {
@@ -1320,66 +553,6 @@ func (op *OperatorConfig) validate() error {
 	for _, login := range op.GitHubLogins {
 		if !githubLoginRe.MatchString(login) {
 			return fmt.Errorf("operator.github_logins: %q is not a valid GitHub login (^[a-zA-Z0-9-]{1,39}$)", login)
-		}
-	}
-	return nil
-}
-
-// validateExtensions checks extension config for an agent identified by key.
-func (ac *AgentConfig) validateExtensions(key string) error {
-	for _, mp := range ac.Extensions.Marketplaces {
-		if mp.Name == "" || mp.Repo == "" {
-			return fmt.Errorf("agent %s: marketplace entry missing name or repo", key)
-		}
-		if !extensionNameRe.MatchString(mp.Name) {
-			return fmt.Errorf("agent %s: marketplace name %q contains invalid characters", key, mp.Name)
-		}
-		if !extensionRepoRe.MatchString(mp.Repo) {
-			return fmt.Errorf("agent %s: marketplace repo %q is not a valid owner/repo", key, mp.Repo)
-		}
-		if mp.Commit != "" && !commitHashRe.MatchString(mp.Commit) {
-			return fmt.Errorf("agent %s: marketplace commit %q is not a valid hex hash", key, mp.Commit)
-		}
-	}
-	for _, pl := range ac.Extensions.Plugins {
-		if pl.Name == "" || pl.Marketplace == "" {
-			return fmt.Errorf("agent %s: plugin entry missing name or marketplace", key)
-		}
-		if !extensionNameRe.MatchString(pl.Name) {
-			return fmt.Errorf("agent %s: plugin name %q contains invalid characters", key, pl.Name)
-		}
-		if !extensionNameRe.MatchString(pl.Marketplace) {
-			return fmt.Errorf("agent %s: plugin marketplace %q contains invalid characters", key, pl.Marketplace)
-		}
-	}
-	for _, sk := range ac.Extensions.Skills {
-		if sk.Name == "" || sk.Repo == "" {
-			return fmt.Errorf("agent %s: skill entry missing name or repo", key)
-		}
-		if !extensionNameRe.MatchString(sk.Name) {
-			return fmt.Errorf("agent %s: skill name %q contains invalid characters", key, sk.Name)
-		}
-		if !extensionRepoRe.MatchString(sk.Repo) {
-			return fmt.Errorf("agent %s: skill repo %q is not a valid owner/repo", key, sk.Repo)
-		}
-	}
-	for _, mcp := range ac.Extensions.MCPServers {
-		if mcp.Name == "" {
-			return fmt.Errorf("agent %s: mcp_server entry missing name", key)
-		}
-		if mcp.URL == "" && mcp.Command == "" {
-			return fmt.Errorf("agent %s: mcp_server %s requires command or url", key, mcp.Name)
-		}
-		vaultSet := make(map[string]bool, len(ac.Vaults))
-		for _, v := range ac.Vaults {
-			vaultSet[v] = true
-		}
-		for envKey, envVal := range mcp.Env {
-			for _, m := range vaultRefRe.FindAllStringSubmatch(envVal, -1) {
-				if !vaultSet[m[1]] {
-					return fmt.Errorf("agent %s: mcp_server %s: env %s references vault %q not in agent vaults list", key, mcp.Name, envKey, m[1])
-				}
-			}
 		}
 	}
 	return nil
